mgr/cephadm: move _apply_service to serve.py
(cherry picked from commit 6b2664a9e3dd531d99b5cc7f01384298e0bdc055)
